Change to layer model for actors
Instead of using passable to define if an actor can share the
field with another actor, a universal layer model is implemented.
Two actors on the same layer cannot share a field, whereas two actors
on different layers can.

This also allows for "flying" actors as multiple layers can be used